django_base/requirements.txt
django
djangorestframework
gunicorn
requests
